requirements.txt
asgiref==3.2.3
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
choices==0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
coveralls==1.8.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.9
django-basic-jwt-auth==0.2
django-braces==1.13.0
django-cors-headers==3.1.0
django-heroku==0.3.1
django-model-utils==3.2.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docopt==0.6.2
ecdsa==0.15
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.2
filelock==3.0.12
flake8==3.7.8
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.20
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.1
Pillow==6.2.0
pluggy==0.12.0
psycopg2==2.7.7
psycopg2-binary==2.8.3
py==1.8.0
pyasn1==0.4.8
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-jose==3.1.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
simplejson==3.17.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
uritemplate==3.0.1
urllib3==1.25.3
virtualenv==16.7.5
whitenoise==4.1.3
zipp==0.6.0
